refactoring codebase, finalised code to fit architecture still some improvements to be made generally
