+++ lib/app/app.router.dart
@@ -9,32 +9,29 @@
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 
-import '../ui/auth/authInputTestView/authInputTestView.dart';
-import '../ui/auth/check_mail/check_email_view.dart';
-import '../ui/auth/forget_password/reset_password_view.dart';
-import '../ui/auth/forgotPasswordPage/forgot_password_view.dart';
-import '../ui/auth/login/login_view.dart';
-import '../ui/auth/signUp/signup_view.dart';
-import '../ui/main/home/home_view.dart';
-import '../ui/startup/startup_view.dart';
+import '../ui/auth/auth_input/auth_input.dart';
+import '../ui/auth/check_mail_page/check_email_view.dart';
+import '../ui/auth/forgot_password_page/forgot_password_view.dart';
+import '../ui/auth/login_page/login_view.dart';
+import '../ui/auth/sign_up_page/sign_up_view.dart';
+import '../ui/main/home_page/home_view.dart';
+import '../ui/startup_page/startup_view.dart';
 
 class Routes {
   static const String homeView = '/home-view';
-  static const String authInputTestView = '/auth-input-test-view';
+  static const String authInput = '/auth-input';
   static const String loginView = '/';
   static const String startUpView = '/start-up-view';
   static const String signUpView = '/sign-up-view';
-  static const String resetPasswordView = '/reset-password-view';
-  static const String forgotPasswordPage = '/forgot-password-page';
+  static const String forgotPasswordView = '/forgot-password-view';
   static const String checkEmailView = '/check-email-view';
   static const all = <String>{
     homeView,
-    authInputTestView,
+    authInput,
     loginView,
     startUpView,
     signUpView,
-    resetPasswordView,
-    forgotPasswordPage,
+    forgotPasswordView,
     checkEmailView,
   };
 }
@@ -44,12 +41,11 @@ class StackedRouter extends RouterBase {
   List<RouteDef> get routes => _routes;
   final _routes = <RouteDef>[
     RouteDef(Routes.homeView, page: HomeView),
-    RouteDef(Routes.authInputTestView, page: AuthInputTestView),
+    RouteDef(Routes.authInput, page: AuthInput),
     RouteDef(Routes.loginView, page: LoginView),
     RouteDef(Routes.startUpView, page: StartUpView),
     RouteDef(Routes.signUpView, page: SignUpView),
-    RouteDef(Routes.resetPasswordView, page: ResetPasswordView),
-    RouteDef(Routes.forgotPasswordPage, page: ForgotPasswordPage),
+    RouteDef(Routes.forgotPasswordView, page: ForgotPasswordView),
     RouteDef(Routes.checkEmailView, page: CheckEmailView),
   ];
   @override
@@ -61,9 +57,9 @@ class StackedRouter extends RouterBase {
         settings: data,
       );
     },
-    AuthInputTestView: (data) {
+    AuthInput: (data) {
       return MaterialPageRoute<dynamic>(
-        builder: (context) => const AuthInputTestView(),
+        builder: (context) => const AuthInput(),
         settings: data,
       );
     },
@@ -85,15 +81,9 @@ class StackedRouter extends RouterBase {
         settings: data,
       );
     },
-    ResetPasswordView: (data) {
+    ForgotPasswordView: (data) {
       return MaterialPageRoute<dynamic>(
-        builder: (context) => const ResetPasswordView(),
-        settings: data,
-      );
-    },
-    ForgotPasswordPage: (data) {
-      return MaterialPageRoute<dynamic>(
-        builder: (context) => const ForgotPasswordPage(),
+        builder: (context) => const ForgotPasswordView(),
         settings: data,
       );
     },
