+++ lib/ui/auth/sign_up_page/sign_up_view.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:stacked/stacked.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:zc_desktop_flutter/ui/shared/dumb_widgets/appbar/app_bar.dart';
+import 'package:zc_desktop_flutter/ui/shared/dumb_widgets/app_bar/app_bar.dart';
 import 'package:zc_desktop_flutter/ui/shared/dumb_widgets/build_left_startup_image.dart';
 import 'package:zcdesk_ui/zcdesk_ui.dart';
-import 'components/signup_viewmodel.dart';
+import 'sign_up_viewmodel.dart';
 import 'package:zcdesk_ui/src/shared/styles.dart';
 
 class SignUpView extends StatelessWidget {
