+++ lib/ui/main/message_page/message_view.dart
@@ -1,20 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:stacked/stacked.dart';
-import 'package:zc_desktop_flutter/ui/main/profile_dialog/profile_dialog_view.dart';
+import 'package:zc_desktop_flutter/ui/main/profile_modal/profile_modal_view.dart';
 import 'package:zcdesk_ui/zcdesk_ui.dart';
 
 import 'message_viewmodel.dart';
 
-class MessageView extends StatefulWidget {
-  const MessageView({Key? key}) : super(key: key);
-
-  @override
-  State<MessageView> createState() => _MessageViewState();
-}
-
-class _MessageViewState extends State<MessageView> {
-  final ScrollController _controllerOne = ScrollController();
+class MessageView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ViewModelBuilder<MessageViewModel>.reactive(
@@ -25,16 +17,13 @@ class _MessageViewState extends State<MessageView> {
           child: Padding(
             padding: const EdgeInsets.symmetric(horizontal: 12.0),
             child: Scrollbar(
-              controller: _controllerOne,
+              controller: model.controllerOne,
               child: SingleChildScrollView(
-                controller: _controllerOne,
+                controller: model.controllerOne,
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.stretch,
                   children: [
-                    MessageHeader(
-                      userDefaultImageUrl: 'assets/images/mark.jpg',
-                      userDisplayName: model.userDisplayName,
-                    ),
+                    messageHeader(context, model),
                     horizontalSpaceTiny,
                     Padding(
                         padding: const EdgeInsets.only(left: 51),
@@ -51,21 +40,16 @@ class _MessageViewState extends State<MessageView> {
                       padding: const EdgeInsets.only(left: 51.0),
                       child: Row(
                         children: [
-                          MessageReactions(
-                            emojiIconPath: 'assets/images/🤘🏻.png',
-                          ),
+                          messageReactions(model),
                           horizontalSpaceSmall,
-                          MessageReactions(emojiIconPath: 'assets/images/🎉.png'),
+                          messageReactions(model),
                           horizontalSpaceSmall,
-                          MessageReactions(emojiIconPath: 'assets/images/🔥.png'),
+                          messageReactions(model),
                         ],
                       ),
                     ),
                     verticalSpaceMedium,
-                    MessageHeader(
-                      userDefaultImageUrl: 'assets/images/2.png',
-                      userDisplayName: 'Naza',
-                    ),
+                    messageHeader(context, model),
                     horizontalSpaceTiny,
                     Padding(
                         padding: const EdgeInsets.only(left: 51),
@@ -78,10 +62,7 @@ class _MessageViewState extends State<MessageView> {
                             },
                             child: SelectableText(model.userPost))),
                     verticalSpaceMedium,
-                    MessageHeader(
-                      userDefaultImageUrl: 'assets/images/2.png',
-                      userDisplayName: 'Naza',
-                    ),
+                    messageHeader(context, model),
                     horizontalSpaceTiny,
                     Padding(
                         padding: const EdgeInsets.only(left: 51),
@@ -98,21 +79,16 @@ class _MessageViewState extends State<MessageView> {
                       padding: const EdgeInsets.only(left: 51.0),
                       child: Row(
                         children: [
-                          MessageReactions(
-                            emojiIconPath: 'assets/images/🤘🏻.png',
-                          ),
+                          messageReactions(model),
                           horizontalSpaceSmall,
-                          MessageReactions(emojiIconPath: 'assets/images/🎉.png'),
+                          messageReactions(model),
                           horizontalSpaceSmall,
-                          MessageReactions(emojiIconPath: 'assets/images/🔥.png'),
+                          messageReactions(model),
                         ],
                       ),
                     ),
                     verticalSpaceMedium,
-                    MessageHeader(
-                      userDefaultImageUrl: 'assets/images/6.png',
-                      userDisplayName: 'Nonso',
-                    ),
+                    messageHeader(context, model),
                     horizontalSpaceTiny,
                     Padding(
                         padding: const EdgeInsets.only(left: 51),
@@ -127,9 +103,7 @@ class _MessageViewState extends State<MessageView> {
                     verticalSpaceSmall,
                     Padding(
                       padding: const EdgeInsets.only(left: 51.0),
-                      child: MessageReplies(
-                        numberOfReplies: model.numberOfReplies,
-                      ),
+                      child: messageReplies(model),
                     ),
                     verticalSpaceMedium,
                   ],
@@ -145,113 +119,69 @@ class _MessageViewState extends State<MessageView> {
   }
 }
 
-class MessageHeader extends StatelessWidget {
-  final String userDefaultImageUrl;
-  final String userDisplayName;
-  MessageHeader({
-    Key? key,
-    required this.userDefaultImageUrl,
-    required this.userDisplayName,
-  }) : super(key: key);
-  final DateTime currentMessageTime = DateTime.now();
-  @override
-  Widget build(BuildContext context) {
-    return ViewModelBuilder<MessageViewModel>.reactive(
-      builder: (context, model, child) => Row(children: [
-        GestureDetector(
-          onTap: (){
-            showDialog(context: context, builder: (_)=>ProfileDialogView(profileImgUrl: userDefaultImageUrl, userName: userDisplayName,));
-          },
-          child: CircleAvatar(
-            backgroundImage: AssetImage(userDefaultImageUrl),
-            radius: 22,
-          ),
-        ),
-        SizedBox(
-          width: 6.w,
-        ),
-        Text(userDisplayName,
-            style: TextStyle(
-                fontSize: 20.sp, color: headerColor, fontFamily: 'Lato')),
-        SizedBox(
-          width: 6.w,
-        ),
-        Text('${currentMessageTime.hour}:${currentMessageTime.minute}',
-            style:
-                TextStyle(fontSize: 20.sp, color: timeColor, fontFamily: 'Lato')),
-      ]),
-      viewModelBuilder: () => MessageViewModel(),
-    );
-  }
-}
-
-class MessageReactions extends StatefulWidget {
-  final String emojiIconPath;
-
-  const MessageReactions({
-    Key? key,
-    required this.emojiIconPath,
-  }) : super(key: key);
-
-  @override
-  State<MessageReactions> createState() => _MessageReactionsState();
-}
-
-class _MessageReactionsState extends State<MessageReactions> {
-  int _numberOfReactions = 0;
-  void _increaseReactionNumber() {
-    setState(() {
-      _numberOfReactions++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ViewModelBuilder<MessageViewModel>.reactive(
-      builder: (context, model, child) => Column(children: [
-        Row(
-          children: [
-            InkWell(
-              child: Container(
-                //width: 30,
-                decoration: BoxDecoration(
-                    color: kcBackgroundColor1,
-                    borderRadius: BorderRadius.circular(10)),
-                child: Padding(
-                  padding: const EdgeInsets.all(4.0),
-                  child: Row(children: [
-                    Image.asset(widget.emojiIconPath, height: 15, width: 15),
-                    horizontalSpaceTiny,
-                    Text('$_numberOfReactions'),
-                  ]),
-                ),
-              ),
-              onTap: () {
-                print("reaction incremented");
-                _increaseReactionNumber();
-                //increment reaction
-              },
+Widget messageHeader(BuildContext context, MessageViewModel model) {
+  return Container(
+        child: Row(children: [
+          GestureDetector(
+            onTap: (){
+              showDialog(context: context, builder: (_)=>ProfileModalView());
+            },
+            child: CircleAvatar(
+              backgroundImage: AssetImage(model.userDefaultImageUrl),
+              radius: 22,
             ),
-          ],
-        ),
-        verticalSpaceSmall,
-      ]),
-      viewModelBuilder: () => MessageViewModel(),
+          ),
+          SizedBox(
+            width: 6.w,
+          ),
+          Text(model.userDisplayName,
+              style: TextStyle(
+                  fontSize: 20.sp, color: headerColor, fontFamily: 'Lato')),
+          SizedBox(
+            width: 6.w,
+          ),
+          Text('${model.currentMessageTime.hour}:${model.currentMessageTime.minute}',
+              style:
+                  TextStyle(fontSize: 20.sp, color: timeColor, fontFamily: 'Lato')),
+        ]),
     );
   }
-}
 
-class MessageReplies extends StatelessWidget {
-  final int numberOfReplies;
+Widget messageReactions(MessageViewModel model) {
+  return Container(
+        child: Column(children: [
+          Row(
+            children: [
+              InkWell(
+                child: Container(
+                  //width: 30,
+                  decoration: BoxDecoration(
+                      color: kcBackgroundColor1,
+                      borderRadius: BorderRadius.circular(10)),
+                  child: Padding(
+                    padding: const EdgeInsets.all(4.0),
+                    child: Row(children: [
+                      Image.asset(model.emojiIconPath, height: 15, width: 15),
+                      horizontalSpaceTiny,
+                      Text('${model.numberOfReactions}'),
+                    ]),
+                  ),
+                ),
+                onTap: () {
+                  print("reaction incremented");
+                  model.increaseReactionNumber();
+                  //increment reaction
+                },
+              ),
+            ],
+          ),
+          verticalSpaceSmall,
+        ]),
+    );
+  }
 
-  const MessageReplies({
-    Key? key,
-    required this.numberOfReplies,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
+Widget messageReplies(MessageViewModel model) {
+  return InkWell(
       onTap: () {
         print("Threads opened");
       },
@@ -266,7 +196,7 @@ class MessageReplies extends StatelessWidget {
             constructRepliesHighLightAvatars(
                 userDefaultImageUrl: 'assets/images/7.png'),
             constructRepliesHighLightAvatars(
-                userDefaultImageUrl: MessageViewModel().userDefaultImageUrl),
+                userDefaultImageUrl: model.userDefaultImageUrl),
             horizontalSpaceTiny,
             //TODO: Get message time of last reply in a thread
             Text('${MessageViewModel().numberOfReplies} replies',
@@ -278,7 +208,6 @@ class MessageReplies extends StatelessWidget {
       ),
     );
   }
-}
 
 Widget constructRepliesHighLightAvatars({
   required String userDefaultImageUrl,
