+++ lib/ui/main/message_page/message_viewmodel.dart
@@ -1,21 +1,26 @@
+import 'package:flutter/cupertino.dart';
 import 'package:stacked/stacked.dart';
 
 import 'package:zc_desktop_flutter/app/app.logger.dart';
 
 class MessageViewModel extends BaseViewModel {
-  final log = getLogger("MessageViewwModel");
+  final log = getLogger("MessageViewModel");
 
-  String _userDefaultImageUrl = 'assets/images/zc_logo.png';
-  int _numberOfReplies = 14;
-  String _userDisplayName = 'Mark';
+  final DateTime currentMessageTime = DateTime.now();
+  final ScrollController controllerOne = ScrollController();
 
-  String _emojiIcon = 'assets/images/🤘🏻.png';
-  final String _userPost =
+  String userDefaultImageUrl = 'assets/images/zc_logo.png';
+  int numberOfReplies = 14;
+  int numberOfReactions = 0;
+  String userDisplayName = 'Mark';
+
+  String emojiIconPath = 'assets/images/🤘🏻.png';
+  String userPost =
       'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.';
 
-  String get userDefaultImageUrl => _userDefaultImageUrl;
-  int get numberOfReplies => _numberOfReplies;
-  String get userDisplayName => _userDisplayName;
-  String get userPost => _userPost;
-  String get emojiIcon => _emojiIcon;
+
+  void increaseReactionNumber() {
+    numberOfReactions++;
+    notifyListeners();
+  }
 }
