+++ lib/ui/main/profile_modal/profile_modal_viewmodel.dart
@@ -2,8 +2,12 @@ import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 import 'package:zc_desktop_flutter/app/app.locator.dart';
 
-class ProfileDialogViewModel extends BaseViewModel {
+class ProfileModalViewModel extends BaseViewModel {
   final _navigationService = locator<NavigationService>();
+
+  final String profileImgUrl = "";
+  final String userName = "";
+
   bool _fav = false;
   bool get fav => _fav;
