--- lib/ui/shared/dumb_widgets/center_list_tile/center_tile_viewmodel.dart
@@ -1,12 +1,0 @@
-final Map<String, String> members = {
-  'WilliamTobs': 'images/widgetImages/ava3.png',
-  'Geeky_Tobi': 'images/widgetImages/ava.png',
-  'Susan': 'images/widgetImages/ava2.jpg',
-  'Jay': 'images/widgetImages/ava.png',
-  'Jayda': 'images/widgetImages/ava2.jpg',
-  'MJ': 'images/widgetImages/ava3.png',
-  'Raphael': 'images/widgetImages/ava.png',
-  'Jake': 'images/widgetImages/ava.png',
-  'Dan': 'images/widgetImages/ava2.jpg',
-  'Eri': 'images/widgetImages/ava3.png',
-};
