+++ pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   flutter_svg: ^0.22.0
   google_fonts: ^2.1.0
   bitsdojo_window: ^0.1.1+1
+  desktop_window: ^0.4.0
   intl: ^0.17.0
 
   # local storage
